project updated

+++ RentCar/RentCar/Models/Entities/Car.cs
@@ -17,8 +17,8 @@
         public string CreatedByEmail {  get; set; }
         public int FuelCapacity { get; set; }
         public string City { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public string OwnerPhoneNumber { get; set; }
     }
 }
